requirements.txt
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
requests>=2.26.0
beautifulsoup4>=4.9.0
openpyxl>=3.0.0
python-dotenv>=0.21.0
streamlit>=1.32.0 